requirements.txt
pyannote.audio
librosa
matplotlib
ffmpeg-python
pydub
numpy
scipy
soundfile
